+++ lib/brando/plugs/i18n.ex
@@ -3,6 +3,7 @@ defmodule Brando.Plug.I18n do
   A plug for checking i18n
   """
   import Brando.I18n
+  import Brando.Utils, only: [current_user: 1]
 
   @doc """
   Assign current locale.
@@ -32,27 +33,26 @@ defmodule Brando.Plug.I18n do
   end
 
   @doc """
-  Set locale to current_user's language
+  Set locale to current user's language
 
   This sets both Brando.Gettext (the default gettext we use in the backend),
   as well as delegating to the I18n module for setting proper locale for all
   registered gettext modules.
   """
   @spec put_admin_locale(Plug.Conn.t, Keyword.t) :: Plug.Conn.t
-  def put_admin_locale(%{private: %{plug_session: %{"current_user" => current_user}}} = conn, []) do
-    language = Map.get(current_user, :language, Brando.config(:default_admin_language))
+  def put_admin_locale(conn, []) do
+    language =
+      case current_user(conn) do
+        nil ->
+          Brando.config(:default_admin_language)
+        user ->
+          Map.get(user, :language, Brando.config(:default_admin_language))
+      end
 
     # set for default brando backend
     Gettext.put_locale(Brando.Gettext, language)
     Brando.I18n.put_locale_for_all_modules(language)
     conn
-  end
-
-  @doc """
-  Set default language
-  """
-  def put_admin_locale(conn, []) do
-    assign_language(conn, Brando.config(:default_admin_language))
   end
 
   @spec extract_language_from_path(Plug.Conn.t) :: String.t | nil